+++ 05-ScenceManager/Pipe.h
@@ -3,8 +3,8 @@
 class Pipe : public CGameObject
 {
 public:
-	int model;
-	Pipe(int model);
+	float model;
+	Pipe(float model);
 	virtual void Render();
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
 	~Pipe();
