+++ 05-ScenceManager/Portal.cpp
@@ -1,7 +1,7 @@
 #include "Portal.h"
 
 
-CPortal::CPortal(float l, float t, float r, float b, int scene_id )
+CPortal::CPortal(float l, float t, float r, float b, float scene_id )
 {
 	this->scene_id = scene_id;
 	x = l;
